Unify the file names
