+++ src/apps/hatenaBlogApp.ts
