+++ tests/apps/hatenaBookmarkApp.test.ts
@@ -1,9 +1,9 @@
 import { mockSpreadSheet } from '../setupMock';
-import { UCCardApp } from '../../src/apps/ucCardApp';
+import { HatenaBookmarkApp } from '../../src/apps/hatenaBookmarkApp';
 import * as upload from '../../src/upload';
 
-describe('UCCard test', () => {
-  const APP_NAME = 'UCCard';
+describe('HatenaBookmarkApp test', () => {
+  const APP_NAME = 'HatenaBookmark';
   let postSpy: jest.SpyInstance<void, [message?: string]>;
 
   beforeEach(() => {
@@ -19,55 +19,69 @@ describe('UCCard test', () => {
       const sheet = mockSpreadSheet.getSheetByName(APP_NAME);
       expect(sheet.getLastRow()).toBe(0);
 
-      const app = new UCCardApp();
+      const app = new HatenaBookmarkApp();
       app.create();
 
       expect(UrlFetchApp.fetch).toHaveBeenCalledWith(
-        'https://www2.uccard.co.jp/important/'
+        'https://b.hatena.ne.jp/hotentry.rss'
+      );
+
+      expect(UrlFetchApp.fetch).toHaveBeenCalledWith(
+        'https://b.hatena.ne.jp/hotentry/it.rss'
       );
 
       expect(sheet.getLastRow()).toBe(2);
-      expect(sheet.getRange(1, 1).getDisplayValue()).toBe('2020年12月20日');
+      expect(sheet.getRange(1, 1).getDisplayValue()).toBe(
+        '2020-12-28T00:00:00Z'
+      );
       expect(sheet.getRange(1, 2).getDisplayValue()).toBe('タイトル１');
       expect(sheet.getRange(1, 3).getDisplayValue()).toBe(
-        'https://www2.uccard.co.jp/uccard/1'
+        'https://example.com/hatena/1'
       );
-      expect(sheet.getRange(2, 1).getDisplayValue()).toBe('2020年12月19日');
+      expect(sheet.getRange(2, 1).getDisplayValue()).toBe(
+        '2020-12-27T00:00:00Z'
+      );
       expect(sheet.getRange(2, 2).getDisplayValue()).toBe('タイトル２');
       expect(sheet.getRange(2, 3).getDisplayValue()).toBe(
-        'https://www2.uccard.co.jp/uccard/2'
+        'https://example.com/hatena/2'
       );
     });
 
     test('fetch and create infos without duplicate title', async () => {
       const sheet = mockSpreadSheet.getSheetByName(APP_NAME);
       sheet.insertRows(1, 2);
-      sheet.getRange(1, 1).setValue('2020年12月22日');
+      sheet.getRange(1, 1).setValue('2020-12-25T00:00:00Z');
       sheet.getRange(1, 2).setValue('タイトル１');
-      sheet.getRange(1, 3).setValue('https://www2.uccard.co.jp/uccard/1');
-      sheet.getRange(2, 1).setValue('2020年12月24日');
+      sheet.getRange(1, 3).setValue('https://example.com/hatena/1');
+      sheet.getRange(2, 1).setValue('2020-12-25T00:00:00Z');
       sheet.getRange(2, 2).setValue('新タイトル');
-      sheet.getRange(2, 3).setValue('https://www2.uccard.co.jp/uccard/new');
+      sheet.getRange(2, 3).setValue('https://example.com/hatena/new');
       expect(sheet.getLastRow()).toBe(2);
 
-      const app = new UCCardApp();
+      const app = new HatenaBookmarkApp();
       app.create();
 
       expect(sheet.getLastRow()).toBe(3);
-      expect(sheet.getRange(1, 1).getDisplayValue()).toBe('2020年12月19日');
+      expect(sheet.getRange(1, 1).getDisplayValue()).toBe(
+        '2020-12-27T00:00:00Z'
+      );
       expect(sheet.getRange(1, 2).getDisplayValue()).toBe('タイトル２');
       expect(sheet.getRange(1, 3).getDisplayValue()).toBe(
-        'https://www2.uccard.co.jp/uccard/2'
+        'https://example.com/hatena/2'
       );
-      expect(sheet.getRange(2, 1).getDisplayValue()).toBe('2020年12月22日');
+      expect(sheet.getRange(2, 1).getDisplayValue()).toBe(
+        '2020-12-25T00:00:00Z'
+      );
       expect(sheet.getRange(2, 2).getDisplayValue()).toBe('タイトル１');
       expect(sheet.getRange(2, 3).getDisplayValue()).toBe(
-        'https://www2.uccard.co.jp/uccard/1'
+        'https://example.com/hatena/1'
       );
-      expect(sheet.getRange(3, 1).getDisplayValue()).toBe('2020年12月24日');
+      expect(sheet.getRange(3, 1).getDisplayValue()).toBe(
+        '2020-12-25T00:00:00Z'
+      );
       expect(sheet.getRange(3, 2).getDisplayValue()).toBe('新タイトル');
       expect(sheet.getRange(3, 3).getDisplayValue()).toBe(
-        'https://www2.uccard.co.jp/uccard/new'
+        'https://example.com/hatena/new'
       );
     });
   });
@@ -77,16 +91,16 @@ describe('UCCard test', () => {
       const sheet = mockSpreadSheet.getSheetByName(APP_NAME);
       expect(sheet.getLastRow()).toBe(0);
 
-      const app = new UCCardApp();
+      const app = new HatenaBookmarkApp();
       app.create();
       app.upload();
 
       const messages = [
         `【${APP_NAME}】`,
         'タイトル１:',
-        'https://www2.uccard.co.jp/uccard/1',
+        'https://example.com/hatena/1',
         'タイトル２:',
-        'https://www2.uccard.co.jp/uccard/2',
+        'https://example.com/hatena/2',
       ];
       expect(postSpy).toHaveBeenCalledWith(`${messages.join('\n')}`);
     });
@@ -95,7 +109,7 @@ describe('UCCard test', () => {
       const sheet = mockSpreadSheet.getSheetByName(APP_NAME);
       expect(sheet.getLastRow()).toBe(0);
 
-      const app = new UCCardApp();
+      const app = new HatenaBookmarkApp();
       app.upload();
 
       expect(postSpy).not.toHaveBeenCalled();
